Refactor screens directory
